chore(cmd): split FloatingCommandBar god-component into per-file sub-components
Extract 8 inline sub-components and lift resize/geometry state into
a dedicated hook, reducing the orchestrator from 2832 → 1056 lines.

Resolves #412
